+++ R/mapdeck_map_utilities.R
@@ -131,6 +131,7 @@ layerId <- function(
 		, "cesium"
 		, "i3s"
 		, "geojson"
+		, "geohash"
 		, "greatcircle"
 		, "grid"
 		, "h3"